[Rebuild] contentTypeId를 Int로 수정, imageURL을 base64인코딩기반으로 Data 변환

+++ travelPlan/Source/Data/Network/Server/RequestDTO/DestinationScrapToggleRequestDTO.swift
@@ -8,6 +8,6 @@
 import Foundation
 
 struct DestinationScrapToggleRequestDTO: Encodable {
-  let objectId: Int64
+  let objectId: Int
   let forderName: String?
 }
